LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-4-expected.txt

Toggle popover


FAIL Cases where the next focus candidate isn't in the direct parent scope with popovertarget invocation assert_equals: no more focusable elements after the button expected Element node <button id="after"></button> but got Element node <body><div id="no-focus-candidate">
  <button tabindex="0...
FAIL Cases where the next focus candidate isn't in the direct parent scope with command/commandfor invocation assert_equals: no more focusable elements after the button expected Element node <button id="after"></button> but got Element node <body><div id="no-focus-candidate">
  <button tabindex="0...
FAIL Cases where the next focus candidate isn't in the direct parent scope with imperative invocation assert_equals: no more focusable elements after the button expected Element node <button id="after"></button> but got Element node <body><div id="no-focus-candidate">
  <button tabindex="0...
